+++ misc/changelog.md
@@ -6,6 +6,7 @@
 
 - Rework preserving timestamps (#92).
 - Lower bundle size of all builds.
+- Disable downscaling without a dialog when unsupported.
 
 ### Fixed